Refactor CSV file used to import pegasus data to be shared with dashboard
